Working on resolving issues #65, #67, and #68
